RavenDB-14706 Move common implementations of RelationalDatabase stuff to Common directory
