+++ src/Raven.Server/Documents/ETL/Providers/RelationalDatabase/Common/Metrics/RelationalDatabaseReplicationMetrics.cs
@@ -1,7 +1,7 @@
 ﻿using Raven.Server.Utils.Metrics;
 using Sparrow.Json.Parsing;
 
-namespace Raven.Server.Documents.ETL.Providers.RelationalDatabase.Metrics;
+namespace Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common.Metrics;
 
  public sealed class RelationalDatabaseEtlTableMetrics(string tableName)
  {
