+++ src/Raven.Server/Documents/ETL/Providers/RelationalDatabase/Common/TestRelationalDatabaseEtlScript.cs
@@ -2,7 +2,7 @@ using Raven.Client.Documents.Operations.ConnectionStrings;
 using Raven.Client.Documents.Operations.ETL;
 using Raven.Server.Documents.ETL.Test;
 
-namespace Raven.Server.Documents.ETL.Providers.RelationalDatabase;
+namespace Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common;
 
 public sealed class TestRelationalDatabaseEtlScript<TRelationalConnectionString, TRelationalEtlConfiguration> : TestEtlScript<TRelationalEtlConfiguration, TRelationalConnectionString>
     where TRelationalConnectionString: ConnectionString
