+++ src/Raven.Server/Documents/ETL/Providers/SQL/RelationalWriters/DbProviderFactories.cs
@@ -4,7 +4,7 @@ using Microsoft.Data.SqlClient;
 using Npgsql;
 using Oracle.ManagedDataAccess.Client;
 using Raven.Client.Documents.Operations.ETL.SQL;
-using DbCommandBuilder = Raven.Server.Documents.ETL.Providers.RelationalDatabase.RelationalWriters.DbCommandBuilder;
+using DbCommandBuilder = Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common.RelationalWriters.DbCommandBuilder;
 
 namespace Raven.Server.Documents.ETL.Providers.SQL.RelationalWriters
 {
