+++ src/Raven.Server/Documents/ETL/Providers/Snowflake/Handlers/Processors/SnowflakeEtlHandlerProcessorForTest.cs
@@ -2,6 +2,7 @@
 using Raven.Client.Documents.Operations.ETL.Snowflake;
 using Raven.Server.Documents.ETL.Providers.Raven.Handlers.Processors;
 using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common;
 using Raven.Server.Json;
 using Sparrow.Json;
