+++ src/Raven.Server/Documents/ETL/Providers/Snowflake/RelationalWriters/SnowflakeDatabaseWriter.cs
@@ -4,13 +4,14 @@ using System.Data.Common;
 using Raven.Client.Documents.Operations.ETL;
 using Raven.Client.Documents.Operations.ETL.Snowflake;
 using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
-using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Metrics;
-using Raven.Server.Documents.ETL.Providers.RelationalDatabase.RelationalWriters;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common.Metrics;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common.RelationalWriters;
 using Raven.Server.NotificationCenter.Notifications;
 using Raven.Server.NotificationCenter.Notifications.Details;
 using Snowflake.Data.Client;
 using Sparrow.Json;
-using DbCommandBuilder = Raven.Server.Documents.ETL.Providers.RelationalDatabase.RelationalWriters.DbCommandBuilder;
+using DbCommandBuilder = Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common.RelationalWriters.DbCommandBuilder;
 
 namespace Raven.Server.Documents.ETL.Providers.Snowflake.RelationalWriters;
