+++ src/Raven.Server/Documents/ETL/Providers/Snowflake/SnowflakeDocumentTransformer.cs
@@ -2,10 +2,9 @@
 using System.Linq;
 using Raven.Client.Documents.Operations.ETL;
 using Raven.Client.Documents.Operations.ETL.Snowflake;
-using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common;
 using Raven.Server.Documents.Patch;
 using Raven.Server.ServerWide.Context;
-using Sparrow.Json;
 
 namespace Raven.Server.Documents.ETL.Providers.Snowflake;
 
@@ -19,7 +18,7 @@ internal sealed class SnowflakeDocumentTransformer(
 {
     protected override List<RelationalDatabaseTableWithRecords> GetEtlTables()
     {
-        return _config.SnowflakeTables.Select(RelationalDatabaseTableWithRecords.FromSnowflakeEtlTable).ToList();
+        return Config.SnowflakeTables.Select(RelationalDatabaseTableWithRecords.FromSnowflakeEtlTable).ToList();
     }
 }
