+++ test/SlowTests/Server/Documents/ETL/SQL/SqlEtlTests.cs
@@ -23,7 +23,8 @@ using Raven.Client.Documents.Operations.ETL.SQL;
 using Raven.Client.Extensions;
 using Raven.Server.Config;
 using Raven.Server.Documents.ETL.Providers.RelationalDatabase;
-using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Test;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common;
+using Raven.Server.Documents.ETL.Providers.RelationalDatabase.Common.Test;
 using Raven.Server.Documents.ETL.Providers.SQL;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.SqlMigration;
